add base auth

+++ courses/fields.py
@@ -2,11 +2,12 @@ from django.db import models
 from django.core.exceptions import ObjectDoesNotExist
 
 
-class OrderField(models.PositiveIntegerField):
+class OrderField(models.PositiveBigIntegerField):
     def __init__(self, for_fields=None, *args, **kwargs):
-        self.for_fields = for_fields
+        self.for_fields = for_fields 
         super().__init__(*args, **kwargs)
 
+    
     def pre_save(self, model_instance, add):
         if getattr(model_instance, self.attname) is None:
             try:
@@ -16,10 +17,10 @@ class OrderField(models.PositiveIntegerField):
                     for field in self.for_fields}
                     qs = qs.filter(**query)
                 last_item = qs.latest(self.attname)
-                value = last_item.order + 1
+                value = last_item.order + 1 
             except ObjectDoesNotExist:
-                value = 0
+                value = 0 
             setattr(model_instance, self.attname, value)
-            return value
+            return value 
         else:
             return super().pre_save(model_instance, add)
